+++ lib/src/codec.dart
@@ -2,18 +2,18 @@ library jwt.codec;
 
 import 'dart:convert';
 
+import 'jose.dart';
 import 'jws.dart';
 import 'jwt.dart';
 import 'jwt_claimset.dart';
 
-typedef JsonWebToken<CS> JwtTokenDecoder<CS extends JwtClaimSet>(
-    String jwtToken,
-    {JwsValidationContext validationContext});
+typedef JsonWebToken<H, CS> JwtTokenDecoder<H extends JoseHeader,
+    CS extends JwtClaimSet>(String jwtToken, {JwsValidationContext validationContext});
 
 typedef CS JwtClaimSetDecoder<CS extends JwtClaimSet>(Map claimSetJson,
     {JwsValidationContext validationContext});
 
-JwtTokenDecoder<JwtClaimSet> defaultJwtTokenDecoder(
+JwtTokenDecoder<JoseHeader, JwtClaimSet> defaultJwtTokenDecoder(
     JwtClaimSetDecoder claimSetDecoder) {
   return (String jwtToken, {JwsValidationContext validationContext}) =>
       new JsonWebToken.decode(jwtToken,
@@ -21,13 +21,15 @@ JwtTokenDecoder<JwtClaimSet> defaultJwtTokenDecoder(
           claimSetParser: claimSetDecoder);
 }
 
-class JwtCodec<CS extends JwtClaimSet> extends Codec<JsonWebToken<CS>, String> {
-  final Converter<JsonWebToken<CS>, String> encoder = new JwtEncoder<CS>();
-  final Converter<String, JsonWebToken<CS>> decoder;
+class JwtCodec<H extends JoseHeader, CS extends JwtClaimSet>
+    extends Codec<JsonWebToken<H, CS>, String> {
+  final Converter<JsonWebToken<H, CS>, String> encoder =
+      new JwtEncoder<H, CS>();
+  final Converter<String, JsonWebToken<H, CS>> decoder;
 
   JwtCodec(this.decoder);
 
-  JwtCodec.simple(JwtTokenDecoder<CS> decoder,
+  JwtCodec.simple(JwtTokenDecoder<H, CS> decoder,
       {JwsValidationContextFactory contextFactory})
       : this(new JwtDecoder(decoder, contextFactory));
 
@@ -36,21 +38,21 @@ class JwtCodec<CS extends JwtClaimSet> extends Codec<JsonWebToken<CS>, String> {
       : this(new JwtDecoder(defaultJwtTokenDecoder(decoder), contextFactory));
 }
 
-class JwtDecoder<CS extends JwtClaimSet>
-    extends Converter<String, JsonWebToken<CS>> {
-  final JwtTokenDecoder<CS> decoder;
+class JwtDecoder<H extends JoseHeader, CS extends JwtClaimSet>
+    extends Converter<String, JsonWebToken<H, CS>> {
+  final JwtTokenDecoder<H, CS> decoder;
   final JwsValidationContextFactory contextFactory;
 
   JwtDecoder(this.decoder, this.contextFactory);
 
   @override
-  JsonWebToken<CS> convert(String input) => contextFactory != null
+  JsonWebToken<H, CS> convert(String input) => contextFactory != null
       ? decoder(input, validationContext: contextFactory())
       : decoder(input);
 }
 
-class JwtEncoder<CS extends JwtClaimSet>
-    extends Converter<JsonWebToken<CS>, String> {
+class JwtEncoder<H extends JoseHeader, CS extends JwtClaimSet>
+    extends Converter<JsonWebToken<H, CS>, String> {
   @override
-  String convert(JsonWebToken<CS> input) => input.encode();
+  String convert(JsonWebToken<H, CS> input) => input.encode();
 }